pyproject.toml
[project]
name = "the_likes"
version = "0.1.0"
description = "The Likes app. Find people based on same values."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=4.0.1",
    "alembic>=1.16.4",
    "apscheduler>=3.11.0",
    "async-lru>=2.0.5",
    "asyncpg>=0.30.0",
    "celery[redis]>=5.5.3",
    "click==8.2.1",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[standard]>=0.116.1",
    "geoalchemy2[shapely]>=0.18.0",
    "geopy>=2.4.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pwnedpasswords>=2.0.0",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "sqlalchemy>=2.0.41",
    "typer>=0.16.0",
    "uvicorn[standard]>=0.35.0",
    "zxcvbn-python>=4.4.24",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "jupyter>=1.1.1",
    "random-password-generator>=2.2.0",
    "ruff>=0.12.7",
]

[tool.ruff]
line-length = 79
unfixable = ["F401"]  # unused imports

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]

[tool.ruff.format]
quote-style = "single"
